docs: trim root README
