+++ alita_agent_prototype/tests/test_manager_agent.py
@@ -9,6 +9,30 @@ def test_manager_agent_process_task(tmp_path):
     config = AlitaConfig(workspace_dir=str(tmp_path))
     manager = ManagerAgent(config)
 
+    async def mock_gen(name, desc, ctx):
+        return f"""#!/usr/bin/env python3
+import sys
+import json
+
+def execute(params: dict):
+    result_data = {{
+        'status': 'success',
+        'message': f"Tool '{name}' executed successfully with mock logic.",
+        'received_params': params,
+    }}
+    return result_data
+
+if __name__ == '__main__':
+    if not sys.stdin.isatty():
+        input_params = json.load(sys.stdin)
+        result = execute(input_params)
+        print(json.dumps(result, indent=2))
+    else:
+        print('This script should be run with JSON input via stdin.')
+"""
+
+    manager.mcp_system.llm_code_generator = mock_gen
+
     async def run():
         query = "echo test"
         result = await manager.process_task(query)
@@ -25,7 +49,26 @@ def test_manager_agent_process_task_default():
 
     # Patch MCPSystem generator to avoid real API calls
     async def mock_gen(name, desc, ctx):
-        return agent.mcp_system._mock_llm_code_generation(name, desc, ctx)
+        return f"""#!/usr/bin/env python3
+import sys
+import json
+
+def execute(params: dict):
+    result_data = {{
+        'status': 'success',
+        'message': f"Tool '{name}' executed successfully with mock logic.",
+        'received_params': params,
+    }}
+    return result_data
+
+if __name__ == '__main__':
+    if not sys.stdin.isatty():
+        input_params = json.load(sys.stdin)
+        result = execute(input_params)
+        print(json.dumps(result, indent=2))
+    else:
+        print('This script should be run with JSON input via stdin.')
+"""
     agent.mcp_system.llm_code_generator = mock_gen
 
     async def run():
